Refactors countEntriesInWeek function
Now utilizes findEntryIdx to see if a given date resides in the logged
entries. Simply start from startWeek and iterate over the days up to
endWeek, where each iteration a call to findEntryIdx is made. This
solution is much more robust by not returning right away if stardWeek
does not show up in the logged entries.