Refactors Macros and UserInfo struct. Refactors ReadConfig and saveUserInfo functions. Updates existing function that worked with yaml. Refactors active field for the PhaseInfo struct. config file to work with database tables.
Macros and UserInfo structs now have an id field that comes from the
database table.

saveUserInfo function allows the user's information to be stored in the
database format. saveUserInfo make three function calls, where each
call is made to save the state of the UserInfo, PhaseInfo, and Macros
structs. UpdateUserInfo now takes a database connection as an argument
to pass to saveUserInfo function call and commits transaction
upon returning.

The Active field for the PhaseInfo struct is now replaced with a Status
field that must be one of the stored values in the corresponding
database table: active, stopped, scheduled, or paused.