Adds fields to the Food struct. Gets updated brand name and food price from the user when updating a food. Simplifies getFoodNutrientsUserInput function. Adds UpdateFood function.
Removes getting calorie information from the user in the
getFoodNutrientsUserInput function. The user entered macros are now
utilized to calculate the food calories. calculateCalories is a helper
function for performing the calculation.

UpdateFood adds a vital function for updating fields in the foods table.
This function allows fields like the price of the food to only reside in
the foods table. Allows allows for quality of life changes such as being
able to shorten the name of a food with a really long name. This entails
getting main food information and food nutrients and making updates to
relevant tables.