Lets user log multiple foods and lets user back out of logging a food
Users are now able to log multiple foods for a single session, using the
same date for food entries.

When logging a food there isn't information on any existing food
preferences or suggested serving size. This typically provides insight
into whether or not this is the particular food you were searching for.
Users now have the option to inspect a particular food without logging
it.