feat(search): support meal creation, meal food deletion, editing, and adding.
Meal creation is performed by pressing <ctrl+enter> while focused on the
input field to used the input field text as the name for a new meal.

Adding a food to an existing meal is intuitive. While highlighting a
meal name, press <a> to bring up a new search ui to select a food. Once
selected, add food to the meal.

Individual meal foods serving size and number of servings can be edited
by pressing <e> on highlighed meal food to bring up a editing form.

Deletion is performed through and initial key press, followed by a pop-up
confirmation form.